+++ i2c.h
@@ -54,14 +54,14 @@ typedef enum __attribute__ ((__packed__))
 
 _Static_assert(sizeof(i2c_error_t) == 1, "sizeof(i2c_error_t) != 1");
 
-i2c_error_t		i2c_init(uint8_t sda_index, uint8_t scl_index);
+i2c_error_t		i2c_init(unsigned int sda_index, unsigned int scl_index);
 i2c_error_t		i2c_reset(void);
-i2c_error_t		i2c_send(uint8_t address, uint16_t length, const uint8_t *bytes);
-i2c_error_t		i2c_receive(uint8_t address, uint16_t length, uint8_t *bytes);
-uint16_t		i2c_error_format_string(const char *tag, i2c_error_t error,
-					uint16_t size, char *dst);
+i2c_error_t		i2c_send(unsigned int address, unsigned int length, const uint8_t *bytes);
+i2c_error_t		i2c_receive(unsigned int address, unsigned int length, uint8_t *bytes);
+unsigned int	i2c_error_format_string(const char *tag, i2c_error_t error,
+					unsigned int size, char *dst);
 
-i2c_error_t		i2c_send_1(uint8_t address, uint8_t byte0);
-i2c_error_t		i2c_send_2(uint8_t address, uint8_t byte0, uint8_t byte1);
-i2c_error_t		i2c_send_3(uint8_t address, uint8_t byte0, uint8_t byte1, uint8_t byte2);
+i2c_error_t		i2c_send_1(unsigned int address, unsigned int byte0);
+i2c_error_t		i2c_send_2(unsigned int address, unsigned int byte0, unsigned int byte1);
+i2c_error_t		i2c_send_3(unsigned int address, unsigned int byte0, unsigned int byte1, unsigned int byte2);
 #endif
