+++ util.h
@@ -53,8 +53,8 @@ void reset(void);
 const char *yesno(bool_t value);
 const char *onoff(bool_t value);
 int dprintf(const char *fmt, ...);
-void msleep(uint16_t);
-uint16_t double_to_string(double value, uint8_t precision, double top_decimal, uint16_t size, char *dst);
+void msleep(unsigned int);
+unsigned int double_to_string(double value, unsigned int precision, double top_decimal, unsigned int size, char *dst);
 double string_to_double(const char *);
 
 #endif
